Add 'preferipv4' to the [node] section of tahoe.cfg to prefer ipv4 behavior over other behavior.  I'm not sure if this works yet, but the testing continue.  There may be an issue with how my computer is configured.  It may also be nessecary to include something like this on platforms that only listen on either ipv4 or ipv6, unlike Linux which treats any IPv6 listener as both.
Fix to_human_encoding and OPTIONALHTTPLEAD to be correct.
to_human_encoding is now two functions, one for each version of
IP.

Update URI tests to include ipv6 human encoding tests.  This fixes
many of the tests that failed.  I'm still working on figuring out
if there are more broken tests.

Update the web root.py to check if an address is IPv6 and pretty
print it in the new format for furls.  Update the IPv4 printing
to print it in the new format also.